server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "server for bookmarking app"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "flask>=3.1.1",
    "pinecone>=7.3.0",
    "playwright>=1.54.0",
    "python-dotenv>=1.1.1",
    "supabase>=2.18.1",
]
